refactor: move out components
