+++ packages/js/src/bulk-editor/components/bulk-action-bar.js
@@ -1,82 +1,14 @@
 import CheckIcon from "@heroicons/react/outline/CheckIcon";
-import ChevronDownIcon from "@heroicons/react/outline/ChevronDownIcon";
 import XIcon from "@heroicons/react/outline/XIcon";
 import SolidXIcon from "@heroicons/react/solid/XIcon";
 import { Slot } from "@wordpress/components";
-import { useEffect, useId, useMemo, useRef } from "@wordpress/element";
-import { applyFilters } from "@wordpress/hooks";
+import { useEffect, useId, useRef } from "@wordpress/element";
 import { __, _n, sprintf } from "@wordpress/i18n";
-import { Alert, Button, Checkbox, DropdownMenu, useSvgAria, useToggleState } from "@yoast/ui-library";
-import { BULK_ACTIONS_SLOT, BULK_NOTICES_SLOT, SELECT_MENU_ITEMS_FILTER } from "../constants";
+import { Alert, Button, Checkbox, useSvgAria, useToggleState } from "@yoast/ui-library";
+import { BULK_ACTIONS_SLOT, BULK_NOTICES_SLOT } from "../constants";
 import { useAiUpsell } from "../hooks/use-ai-upsell";
 import { UpsellModal } from "./upsell-modal";
-
-/**
- * A single Select-menu item: an optional icon followed by the label.
- *
- * @param {Object} props      The props.
- * @param {Object} props.item The item ({ key, label, ariaLabel, icon, onClick }).
- *
- * @returns {JSX.Element} The menu item.
- */
-const SelectMenuItem = ( { item } ) => (
-	<DropdownMenu.ButtonItem
-		className="yst-flex yst-items-center yst-gap-2 yst-justify-start yst-px-4 yst-py-2 yst-font-normal yst-text-start yst-text-slate-800 hover:!yst-bg-slate-50 focus:!yst-bg-slate-50"
-		onClick={ item.onClick }
-		aria-label={ item.ariaLabel }
-	>
-		{ item.icon }
-		{ item.label }
-	</DropdownMenu.ButtonItem>
-);
-
-/**
- * Renders a Select-menu item; the `.map()` callback for the default and smart-select item lists.
- *
- * @param {Object} item The item ({ key, label, ariaLabel, icon, onClick }).
- *
- * @returns {JSX.Element} The menu item.
- */
-const renderItem = ( item ) => <SelectMenuItem key={ item.key } item={ item } />;
-
-/**
- * The "Select" menu.
- *
- * @param {Object}   props                  The props.
- * @param {Function} props.onSelectAll      Selects every row.
- * @param {Function} props.onDeselectAll    Clears the selection.
- * @param {number}   props.selectedCount    The number of selected rows.
- * @param {number}   props.totalCount       The total number of rows.
- * @param {Object[]} [props.smartSelectItems] The select improvement items.
- *
- * @returns {JSX.Element} The select menu.
- */
-const SelectMenu = ( { onSelectAll, onDeselectAll, selectedCount, totalCount, smartSelectItems = [] } ) => {
-	const svgAriaProps = useSvgAria();
-	const defaultItems = useMemo( () => [
-		{ key: "select-all", label: __( "Select all", "wordpress-seo" ), onClick: onSelectAll },
-		{ key: "deselect-all", label: __( "Deselect all", "wordpress-seo" ), onClick: onDeselectAll },
-	], [ onSelectAll, onDeselectAll ] );
-
-	const items = applyFilters( SELECT_MENU_ITEMS_FILTER, defaultItems, { selectedCount, totalCount } );
-
-	return (
-		<DropdownMenu as="div" className="yst-relative">
-			<DropdownMenu.Trigger as={ Button } variant="primary" size="small" className="yst-gap-1.5">
-				{ __( "Select", "wordpress-seo" ) }
-				<ChevronDownIcon className="yst-h-4 yst-w-4" { ...svgAriaProps } />
-			</DropdownMenu.Trigger>
-			{ /* 169px is the narrowest width at which the longest label ("Social descriptions") stays on one line */ }
-			<DropdownMenu.List className="yst-absolute yst-z-10 yst-start-0 yst-top-full yst-mt-1 yst-w-[169px]">
-				{ items.map( renderItem ) }
-				{ smartSelectItems.length > 0 && (
-					<div role="separator" className="yst-my-1 yst-border-t yst-border-slate-200" />
-				) }
-				{ smartSelectItems.map( renderItem ) }
-			</DropdownMenu.List>
-		</DropdownMenu>
-	);
-};
+import { SelectMenu } from "./select-menu";
 
 /**
  * The first toolbar row: the multiselection checkbox, the Select menu and the selected-count.
